Provide a maximum size on the number of tests to run in one group.
The implementation of thread local storage on Android does not free slots
in a fixed size datastructure with 256 slots. When a large number of tests
run in a single process it is easy to exceed the limit. We don't have
control over how many slots are burnt per test, so we just assume that
one slot is burnt per test. We split the tests so only 256 run in a given
process.

BUG=354405

Review URL: https://codereview.chromium.org/453953003

Cr-Original-Commit-Position: refs/heads/master@{#289386}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf22e8e9cc92ea96eb6fb077c65d3986028e6578