A more robust way to set/query CPU properties on devices
Provides a method _ForEachCpu which runs a specified operation on
each CPU file on the device. It is efficient, as all of the commands
are executed using a single shell script on the device (as it was
already done for some of the methods); and now also more robust, as we
get the exit status of each individual command run.

BUG=433056

Review URL: https://codereview.chromium.org/732473003

Cr-Original-Commit-Position: refs/heads/master@{#304927}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18689623bac72514040e422003a89f230549252a