Fix android resources in gn builds
Dependent libraries need to be passed in the reverse order that
build_utils.GetSortedTransitiveDependencies returns.

Also, a multi-library resource zip in package_libraries should be sorted
on the numeric values of the directories in the zip, not the string
values.

Review URL: https://codereview.chromium.org/1007043002

Cr-Original-Commit-Position: refs/heads/master@{#320802}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 46ba0c2ba3d041ee8dac492be81904ee073f4ed8