Fix installation of libgl1-mesa-dev and mesa-common-dev in install-build-deps.sh
On a fresh install of Ubuntu Trusty + Cinnamon, install-build-deps.sh failed, because:

* it was looking for libgl1-mesa-dev
* Cinnamon forced libgl1-mesa-dev-lts-utopic to be installed
* there cannot be two versions of libgl1-mesa-dev* installed

This CL adds libgl1-mesa-dev and mesa-common-dev to the group of packages, for which the correct "mesa_variant" suffix is guessed. It also adds "-lts-utopic" to the list of explored suffixes.

BUG=466787

Review URL: https://codereview.chromium.org/1003853003

Cr-Original-Commit-Position: refs/heads/master@{#320889}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f1e8b17f4827ad6030c19f71ebf8e34c91248e2b