Enable forcing `/std:c++17` via GN argument
Use a new `std_cpp17` argument in Windows builds to enable forcing Win32
MSVC builds to use `/std:c++17`. This is already forced for UWP, but is
needed for Win32 if a module compiled with C++17 needs to link against
the `webrtc.lib` compiled from this repository. Otherwise the layout
mismatch between `absl::optional` (non-c++17) and `std::optional`
(c++17) produces a link error.